alx_travel_app/requirements.txt
django
djangorestframework
django-cors-headers
drf-yasg
celery
mysql-connector-python
django-environ